ml-trainer/requirements.txt
SQLAlchemy==2.0.36
psycopg2-binary
pandas
mlflow
scikit-learn